Moving ComponentType enum to fix import order problems causing build and tests to fail
