+++ src/features/pdf/PDFView.tsx
@@ -7,8 +7,8 @@ import { SummaryComponent } from 'src/components/summary/SummaryComponent';
 import { DisplayGroupContainer } from 'src/features/form/containers/DisplayGroupContainer';
 import { PDF_LAYOUT_NAME } from 'src/features/pdf/data/pdfSlice';
 import classes from 'src/features/pdf/PDFView.module.css';
-import { ComponentType } from 'src/layout';
 import { GenericComponent } from 'src/layout/GenericComponent';
+import { ComponentType } from 'src/layout/LayoutComponent';
 import { ReadyForPrint } from 'src/shared/components/ReadyForPrint';
 import { useExprContext } from 'src/utils/layout/ExprContext';
 import type { LayoutNode } from 'src/utils/layout/hierarchy';
