+++ src/layout/index.ts
@@ -123,17 +123,6 @@ export const FormComponentContext = createContext<IFormComponentContext>({
   baseComponentId: undefined,
 });
 
-/**
- * This enum is used to distinguish purely presentational components
- * from interactive form components that can have formData etc.
- */
-export enum ComponentType {
-  Presentation = 'presentation',
-  Form = 'form',
-  Action = 'action',
-  Container = 'container',
-}
-
 export function getLayoutComponentObject<T extends keyof ComponentClassMap>(type: T): ComponentClassMap[T] {
   if (type && type in components) {
     return components[type as keyof typeof components] as any;
